rename: BookmarkList 를 BookmarkGroupList 로 변경, BookmarkGroup 을 BookmarkGroupCard 로 변경
