+++ src/components/Profile/BookmarkGroupList.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import BookmarkGroup from "./BookmarkGroup";
+import BookmarkGroupCard from "./BookmarkGroupCard";
 import type { BookmarkGroupType } from "@/types/bookmark";
 import { mockBookmarkGroups } from "@/mock/bookmarkGroup.mock";
 
@@ -10,7 +10,7 @@ export default function BookmarkList() {
   return (
     <div className="space-y-1">
       {groups.map((group) => (
-        <BookmarkGroup key={group.groupId} group={group} />
+        <BookmarkGroupCard key={group.groupId} group={group} />
       ))}
     </div>
   );
